requirements.txt
mysql-connector-python
python-dotenv
gspread
requests
pydantic